+++ CHANGELOG.md
@@ -64,6 +64,16 @@ Bugs
 - Fix JSON Parse error for all documents, when setting bulk with previous data being undefined
 ---- Error: Unexpected token u in JSON at position 0
 - Fix Clothing Crafting Removing Wrong Items
+- Fix Various rmlui Menu Bugs
+
+attachment editor plugin
+- Added an attachment editor plugin
+- Command '/attachmenteditor'
+
+rmlui Menu
+- Added new text 'input' option
+- Added various fixes
+- Fixed controls
 
 ColShapes
 - Added onEnterCallback to the interface. Triggers when player enters the ColShape.